+++ unstructured/ingest/main.py
@@ -7,6 +7,7 @@ import sys
 import click
 
 from unstructured.ingest.connector.s3_connector import S3Connector, SimpleS3Config
+from unstructured.ingest.connector.reddit import RedditConnector, SimpleRedditConfig
 from unstructured.ingest.doc_processor.generalized import initialize, process_document
 
 
@@ -20,6 +21,7 @@ class MainProcess:
 
     def initialize(self):
         """Slower initialization things: check connections, load things into memory, etc."""
+        self.doc_connector.initialize()
         initialize()
 
     def cleanup(self):
@@ -78,6 +80,38 @@ class MainProcess:
     help="Connect to s3 without local AWS credentials.",
 )
 @click.option(
+    "--subreddit-name",
+    default=None,
+    help='The name of a subreddit, without the "r\\", e.g. "machinelearning"',
+)
+@click.option(
+    "--reddit-client-id",
+    default=None,
+    help="The client ID, see "
+    "https://praw.readthedocs.io/en/stable/getting_started/quick_start.html#prerequisites"
+    " for more information.",
+)
+@click.option(
+    "--reddit-client-secret",
+    default=None,
+    help="The client secret, see "
+    "https://praw.readthedocs.io/en/stable/getting_started/quick_start.html#prerequisites"
+    " for more information.",
+)
+@click.option(
+    "--reddit-user-agent",
+    default="Unstructured Ingest Subreddit fetcher",
+    help="The user agent to use on the Reddit API, see "
+    "https://praw.readthedocs.io/en/stable/getting_started/quick_start.html#prerequisites"
+    " for more information.",
+)
+@click.option(
+    "--reddit-search-query",
+    default=None,
+    help="If set, return posts using this query. Otherwise, use hot posts.",
+)
+@click.option("--reddit-num-posts", default=10, help="The number of posts to fetch.")
+@click.option(
     "--re-download/--no-re-download",
     default=False,
     help="Re-download files from s3 even if they are already present in --download-dir.",
@@ -114,6 +148,12 @@ class MainProcess:
 @click.option("-v", "--verbose", is_flag=True, default=False)
 def main(
     s3_url,
+    subreddit_name,
+    reddit_client_id,
+    reddit_client_secret,
+    reddit_user_agent,
+    reddit_search_query,
+    reddit_num_posts,
     re_download,
     download_dir,
     preserve_downloads,
@@ -139,6 +179,23 @@ def main(
                 anonymous=s3_anonymous,
                 re_download=re_download,
                 preserve_downloads=preserve_downloads,
+                verbose=verbose,
+            ),
+        )
+    elif subreddit_name:
+        doc_connector = RedditConnector(  # type: ignore
+            config=SimpleRedditConfig(
+                subreddit_name=subreddit_name,
+                client_id=reddit_client_id,
+                client_secret=reddit_client_secret,
+                user_agent=reddit_user_agent,
+                search_query=reddit_search_query,
+                num_posts=reddit_num_posts,
+                # defaults params:
+                download_dir=download_dir,
+                preserve_downloads=preserve_downloads,
+                output_dir=structured_output_dir,
+                re_download=re_download,
                 verbose=verbose,
             ),
         )